0.7
- 2 часть переделывания таймлайна на систему тиков

+++ Assets/Scripts/Keyframe/KeyframeTimeLine/KeyframeArea.cs
@@ -29,7 +29,7 @@ namespace TimeLine
 
         public void OnSelectTrackObject(TrackObjectData trackObjectData)
         {
-            UpdateArea(trackObjectData.trackObject.BeatDuraction);
+            UpdateArea((float)trackObjectData.trackObject.BeatDuraction);
         }
 
         private void UpdateArea(float duraction)
